refactor(cdc): A45-7 decompose cdc.rs into rebuild and polling submodules
Extract cdc trigger rebuild logic into src/cdc/rebuild.rs and polling
CDC (foreign table + matview) into src/cdc/polling.rs. Re-export all
public symbols from src/cdc.rs to maintain the existing API surface.
This keeps the root cdc.rs at ~3386 lines (from 4259) while improving
module clarity.
